Major rework, reorder flow of setup module: start with specifying header index, data type. Add section that loads datasets in the ../Data and Figures/.. sensor raw data subdirectory and displays the unique column headers at each row index. Timestamp columns are then specified, separate from the parameter columns (also remove mention of dropping time-like columns).
The module takes a look at the headers for datasets in the
data and figures subdirectory for the specified sensor and compiles a 
listing
of unique header names, organized by row. The names are associated with 
the
file they appear in, so if the header naming differs from one file to 
the other,
all of the header labels for a particular index will be noted along with 
the files
they appear in. When the standard ingestion module runs, it parses the 
setup.json
and creates a listing of parameter names that pertain to the particular 
data file
being loaded.

This makes the ingestion process more robust, as before, the
ingestion module assumed that the data header names were consistent 
across all
datasets of the same sensor type. A lot of this process takes place 
behind the
scenes, so the user doesnt have to keep track of which formatting scheme 
is
required for which sensor if a scenario arises where some sensors have 
different
formatting that others of the same make and model.