In draw_scatter(): Change default scatter text position from stricly 'upper_left' to be based on whether the slope exceeds 2.0.
In scatter_plotter(): Set the tick spacing using a while loop to 
increment the size of the tick_spacing by 5 until the number of tick 
divisions is less than or equal to 7. 7 is chosen because it seems like 
a good upper limit on the number of ticks before the grid becomes too 
visually busy.