If the ReferenceSetup class has been run, the setup_data are loaded via the json library's json.loads function for converting the configuration info to a accessible dictionary.
Add additional (optional) parameters to the query_aqs function for 
specifying aqs site id or selecting whether or niot to also query met 
data. By default, assume user has completed the ReferenceSetup method 
and has specified the site AQS id - otherwise require that the AQS id be 
passed directly to the query_aqs function.

Add query_airnow function, similar implementation to query_aqs function 
in that the setup config dictionary is searched first for the latitude 
and longitude of the site to create a bounding box. If the config 
dictionary is not found, use the bbox passed to the function.

For both of these API methods, datasets are parsed and subsequently 
added to the self.data attribute for ReferenceMethod objects.