Update __init__.py
Fix KeyError 'Temp' bug that occurred when loading sensortoolkit for the first time after installation. This error occurred on the first import of the library because the parameter dictionary was written to the app_data folder for first import, but the dictionary was not subsequently loaded into the _param_dict dictionary, leaving it blank "{}". Fix is to remove the else statement for loading the parameter data dictionary so that on first import of the library, the parameter dictionary is written to the app data directory and then loaded in.